+++ pydcs_extensions/fa18ef_tanker/__init__.py
@@ -1,1 +1,0 @@
-from .fa18ef_tanker import *