+++ qt_ui/windows/newgame/QNewGameWizard.py
@@ -93,6 +93,7 @@ class NewGameWizard(QtWidgets.QWizard):
             a6a_intruder=self.field("a6a_intruder"),
             a7e_corsair2=self.field("a7e_corsair2"),
             ea6b_prowler=self.field("ea6b_prowler"),
+            e7a_wedgetail=self.field("e7a_wedgetail"),
             f4bc_phantom=self.field("f4bc_phantom"),
             f15d_baz=self.field("f15d_baz"),
             f_15_idf=self.field("f_15_idf"),
@@ -111,6 +112,7 @@ class NewGameWizard(QtWidgets.QWizard):
             oh_6_vietnamassetpack=self.field("oh_6_vietnamassetpack"),
             uh_60l=self.field("uh_60l"),
             jas39_gripen=self.field("jas39_gripen"),
+            mirage_3=self.field("mirage_3"),
             super_etendard=self.field("super_etendard"),
             sk_60=self.field("sk_60"),
             su15_flagon=self.field("su15_flagon"),
@@ -127,6 +129,7 @@ class NewGameWizard(QtWidgets.QWizard):
             chinesemilitaryassetspack=self.field("chinesemilitaryassetspack"),
             russianmilitaryassetspack=self.field("russianmilitaryassetspack"),
             usamilitaryassetspack=self.field("usamilitaryassetspack"),
+            ukrainemilitaryassetspack=self.field("ukrainemilitaryassetspack"),
         )
 
         blue_faction = self.faction_selection_page.selected_blue_faction