+++ qt_ui/windows/newgame/WizardPages/QGeneratorSettings.py
@@ -96,6 +96,8 @@ class GeneratorOptions(QtWidgets.QWizardPage):
         self.registerField("a7e_corsair2", self.a7e_corsair2)
         self.ea6b_prowler = QtWidgets.QCheckBox()
         self.registerField("ea6b_prowler", self.ea6b_prowler)
+        self.e7a_wedgetail = QtWidgets.QCheckBox()
+        self.registerField("e7a_wedgetail", self.e7a_wedgetail)
         self.hercules = QtWidgets.QCheckBox()
         self.registerField("hercules", self.hercules)
         self.oh_6 = QtWidgets.QCheckBox()
@@ -132,6 +134,8 @@ class GeneratorOptions(QtWidgets.QWizardPage):
         self.registerField("f106_deltadart", self.f106_deltadart)
         self.jas39_gripen = QtWidgets.QCheckBox()
         self.registerField("jas39_gripen", self.jas39_gripen)
+        self.mirage_3 = QtWidgets.QCheckBox()
+        self.registerField("mirage_3", self.mirage_3)
         self.super_etendard = QtWidgets.QCheckBox()
         self.registerField("super_etendard", self.super_etendard)
         self.sk_60 = QtWidgets.QCheckBox()
@@ -166,6 +170,8 @@ class GeneratorOptions(QtWidgets.QWizardPage):
         self.registerField("russianmilitaryassetspack", self.russianmilitaryassetspack)
         self.usamilitaryassetspack = QtWidgets.QCheckBox()
         self.registerField("usamilitaryassetspack", self.usamilitaryassetspack)
+        self.ukrainemilitaryassetspack = QtWidgets.QCheckBox()
+        self.registerField("ukrainemilitaryassetspack", self.ukrainemilitaryassetspack)
 
         modHelpText = QtWidgets.QLabel(
             "<p>Select the mods you have installed. If your chosen factions support them, you'll be able to use these mods in your campaign.</p>"
@@ -197,7 +203,12 @@ class GeneratorOptions(QtWidgets.QWizardPage):
                 "CurrentHill USA Military Assets pack (1.1.5)",
                 self.usamilitaryassetspack,
             ),
+            (
+                "CurrentHill Ukraine Military Assets pack (1.1.1)",
+                self.ukrainemilitaryassetspack,
+            ),
             ("EA-6B Prowler (v2.9.4.102)", self.ea6b_prowler),
+            ("E-7A Wedgetail (AI Only, EW Capable)", self.e7a_wedgetail),
             ("F-100 Super Sabre (v2.7.18.30765 patch 20.10.22)", self.f100_supersabre),
             ("F-104 Starfighter (v2.7.11.222.01)", self.f104_starfighter),
             ("F-105 Thunderchief (v2.7.12.23x)", self.f105_thunderchief),
@@ -209,12 +220,13 @@ class GeneratorOptions(QtWidgets.QWizardPage):
             ("F-4B/C Phantom II (2.8.7.204)", self.f4bc_phantom),
             ("F-84G Thunderjet (v2.5.7.01)", self.f84g_thunderjet),
             ("F9F Panther (v2.8.7.101)", self.f9f_panther),
-            ("F/A-18E/F Super Hornet AI Tanker (version 1.4)", self.fa18ef_tanker),
-            ("F/A-18E/F/G Super Hornet (version 2.3.2)", self.fa_18efg),
+            ("F/A-18E/F/G Super Hornet (version 2.4.2)", self.fa_18efg),
+            ("F/A-18E/F Super Hornet AI Tanker (version 2.4.2)", self.fa18ef_tanker),
             ("Frenchpack (v4.9.1)", self.frenchpack),
             ("High Digit SAMs", self.high_digit_sams),
             ("IDF Assets Pack (v1.1 by IDF Mods Project)", self.irondome),
             ("JAS 39 Gripen (v1.8.5-beta)", self.jas39_gripen),
+            ("Mirage III (v2.5.7.01)", self.mirage_3),
             ("OH-6 Cayuse (v1.2)", self.oh_6),
             ("OH-6 Vietnam Asset Pack (v1.0)", self.oh_6_vietnamassetpack),
             ("OV-10A Bronco", self.ov10a_bronco),
@@ -268,6 +280,7 @@ class GeneratorOptions(QtWidgets.QWizardPage):
         self.a6a_intruder.setChecked(s.get("a6a_intruder", False))
         self.a7e_corsair2.setChecked(s.get("a7e_corsair2", False))
         self.ea6b_prowler.setChecked(s.get("ea6b_prowler", False))
+        self.e7a_wedgetail.setChecked(s.get("e7a_wedgetail", False))
         self.hercules.setChecked(s.get("hercules", False))
         self.uh_60l.setChecked(s.get("uh_60l", False))
         self.f4bc_phantom.setChecked(s.get("f4bc_phantom", False))
@@ -283,6 +296,7 @@ class GeneratorOptions(QtWidgets.QWizardPage):
         self.f106_deltadart.setChecked(s.get("f106_deltadart", False))
         self.jas39_gripen.setChecked(s.get("jas39_gripen", False))
         self.sk_60.setChecked(s.get("sk_60", False))
+        self.mirage_3.setChecked(s.get("mirage_3", False))
         self.super_etendard.setChecked(s.get("super_etendard", False))
         self.su30_flanker_h.setChecked(s.get("su30_flanker_h", False))
         self.su57_felon.setChecked(s.get("su57_felon", False))
@@ -302,3 +316,6 @@ class GeneratorOptions(QtWidgets.QWizardPage):
             s.get("russianmilitaryassetspack", False)
         )
         self.usamilitaryassetspack.setChecked(s.get("usamilitaryassetspack", False))
+        self.ukrainemilitaryassetspack.setChecked(
+            s.get("ukrainemilitaryassetspack", False)
+        )