+++ resources/customized_payloads/Su-24MU.lua
@@ -1,0 +1,215 @@
+local unitPayloads = {
+	["name"] = "Su-24MU",
+	["payloads"] = {
+		[1] = {
+			["displayName"] = "Retribution BAI",
+			["name"] = "Retribution BAI",
+			["pylons"] = {
+				[1] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 8,
+				},
+				[2] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 7,
+				},
+				[3] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 1,
+				},
+				[4] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 2,
+				},
+				[5] = {
+					["CLSID"] = "{16602053-4A12-40A2-B214-AB60D481B20E}",
+					["num"] = 5,
+				},
+				[6] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 6,
+				},
+				[7] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 3,
+				},
+			},
+			["tasks"] = {
+				[1] = 32,
+			},
+		},
+		[2] = {
+			["displayName"] = "Retribution DEAD",
+			["name"] = "Retribution DEAD",
+			["pylons"] = {
+				[1] = {
+					["CLSID"] = "{FBC29BFE-3D24-4C64-B81D-941239D12249}",
+					["num"] = 8,
+				},
+				[2] = {
+					["CLSID"] = "{SU24MU_STORMSHADOW}",
+					["num"] = 7,
+				},
+				[3] = {
+					["CLSID"] = "{FBC29BFE-3D24-4C64-B81D-941239D12249}",
+					["num"] = 1,
+				},
+				[4] = {
+					["CLSID"] = "{SU24MU_STORMSHADOW}",
+					["num"] = 2,
+				},
+				[5] = {
+					["CLSID"] = "{16602053-4A12-40A2-B214-AB60D481B20E}",
+					["num"] = 5,
+				},
+			},
+			["tasks"] = {
+				[1] = 32,
+			},
+		},
+		[3] = {
+			["displayName"] = "Retribution OCA/Runway",
+			["name"] = "Retribution OCA/Runway",
+			["pylons"] = {
+				[1] = {
+					["CLSID"] = "{FBC29BFE-3D24-4C64-B81D-941239D12249}",
+					["num"] = 8,
+				},
+				[2] = {
+					["CLSID"] = "{FBC29BFE-3D24-4C64-B81D-941239D12249}",
+					["num"] = 1,
+				},
+				[3] = {
+					["CLSID"] = "{39821727-F6E2-45B3-B1F0-490CC8921D1E}",
+					["num"] = 7,
+				},
+				[4] = {
+					["CLSID"] = "{16602053-4A12-40A2-B214-AB60D481B20E}",
+					["num"] = 5,
+				},
+				[5] = {
+					["CLSID"] = "{39821727-F6E2-45B3-B1F0-490CC8921D1E}",
+					["num"] = 2,
+				},
+			},
+			["tasks"] = {
+				[1] = 32,
+			},
+		},
+		[4] = {
+			["displayName"] = "Retribution Strike",
+			["name"] = "Retribution Strike",
+			["pylons"] = {
+				[1] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 8,
+				},
+				[2] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 7,
+				},
+				[3] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 1,
+				},
+				[4] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 2,
+				},
+				[5] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 6,
+				},
+				[6] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 3,
+				},
+				[7] = {
+					["CLSID"] = "{16602053-4A12-40A2-B214-AB60D481B20E}",
+					["num"] = 5,
+				},
+			},
+			["tasks"] = {
+				[1] = 32,
+			},
+		},
+		[5] = {
+			["displayName"] = "Retribution OCA/Aircraft",
+			["name"] = "Retribution OCA/Aircraft",
+			["pylons"] = {
+				[1] = {
+					["CLSID"] = "{FBC29BFE-3D24-4C64-B81D-941239D12249}",
+					["num"] = 8,
+				},
+				[2] = {
+					["CLSID"] = "{FBC29BFE-3D24-4C64-B81D-941239D12249}",
+					["num"] = 1,
+				},
+				[3] = {
+					["CLSID"] = "{E2C426E3-8B10-4E09-B733-9CDC26520F48}",
+					["num"] = 7,
+				},
+				[4] = {
+					["CLSID"] = "{16602053-4A12-40A2-B214-AB60D481B20E}",
+					["num"] = 5,
+				},
+				[5] = {
+					["CLSID"] = "{E2C426E3-8B10-4E09-B733-9CDC26520F48}",
+					["num"] = 2,
+				},
+				[6] = {
+					["CLSID"] = "{E2C426E3-8B10-4E09-B733-9CDC26520F48}",
+					["num"] = 6,
+				},
+				[7] = {
+					["CLSID"] = "{E2C426E3-8B10-4E09-B733-9CDC26520F48}",
+					["num"] = 3,
+				},
+			},
+			["tasks"] = {
+				[1] = 32,
+			},
+		},
+		[6] = {
+			["displayName"] = "Retribution CAS",
+			["name"] = "Retribution CAS",
+			["pylons"] = {
+				[1] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 8,
+				},
+				[2] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 7,
+				},
+				[3] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 1,
+				},
+				[4] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 2,
+				},
+				[5] = {
+					["CLSID"] = "{16602053-4A12-40A2-B214-AB60D481B20E}",
+					["num"] = 5,
+				},
+				[6] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 6,
+				},
+				[7] = {
+					["CLSID"] = "{CH_AASM250}",
+					["num"] = 3,
+				},
+			},
+			["tasks"] = {
+				[1] = 32,
+			},
+		},
+	},
+	["tasks"] = {
+	},
+	["unitType"] = "Su-24MU",
+}
+return unitPayloads