+++ changelog.md
@@ -11,6 +11,7 @@
 * **[Modding]** Added OH-6 Cayuse (v1.2) mod support, including the Vietnam Asset Pack v1.0
 * **[Modding]** Added VSN EA-6B Prowler mod support (v2.9.4.102)
 * **[Modding]** Added tripod3 Cold War assets mod support (v1.2)
+* **[Modding]** Added VSN Mirage III mod support (2.5.7.01)
 * **[Campaign Setup]** Allow adjustments to naval TGOs (except carriers) on turn 0
 * **[Campaign Design]** Ability to configure specific carrier names & types in campaign's yaml file 
 * **[Mission Generation]** Ability to inject custom kneeboards
@@ -52,6 +53,9 @@
 * **[Modding]** Updated support for CurrentHill's Swedish Asset Pack (v1.1.0)
 * **[Modding]** Support for CurrentHill's Russian Asset Pack (v1.2.0)
 * **[Modding]** Support for CurrentHill's USA Asset Pack (v1.1.5)
+* **[Modding]** Update CJS Super Hornet to 2.4.2
+* **[Modding]** Support for Cowboy's E-7A Wedgetail mod (Supports EW Script Offensive Jamming)
+* **[Plugins]** Added initial AI support for EW Script 2.0
 
 ## Fixes
 * **[UI/UX]** A-10A flights can be edited again